requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# NLP libraries
nltk==3.8.1
spacy==3.6.0
gensim==4.3.1

# Machine learning
xgboost==1.7.6
lightgbm==4.0.0
imbalanced-learn==0.11.0

# Web framework
Flask==2.3.3
Flask-CORS==4.0.0

# Data visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# Utilities
pyyaml==6.0.1
requests==2.31.0
beautifulsoup4==4.12.2
tqdm==4.66.1
joblib==1.3.2

# Development
pytest==7.4.0
black==23.7.0
flake8==6.1.0

# Optional but recommended
jupyter==1.0.0
notebook==7.0.2
flask-cors>=4.0.0
requests>=2.31.0